Move guardian resource to dedicated pkg
